Refactor:Api분리
